Changed the unit tests to use XUnit rather than NUnit.
